UDP: Add UDP provider framework with common implementation
Simple provider that supports datagram endpoints.  This provider
will be used as the basis to construct a utility provider that
can be used to build other providers, or expand their set of
functionality.

This patch only adds the framework for the UDP provider, along
with the start of the utility provider.  We add common code for
validating the fi_info attributes and capabilities.  And define
the utility provider's wait set.  These serve merely as starting
building blocks for how the UDP provider will be constructed,
along with how it will use the utility provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>